content-security-policy/connect-src/connect-src-websocket-blocked.sub.html

<!DOCTYPE html>
<html>

<head>
    <!-- Programmatically converted from a WebKit Reftest, please forgive resulting idiosyncraciws.-->
    <meta http-equiv="Content-Security-Policy" content="connect-src 'self'; script-src 'self' 'unsafe-inline';">
    <title>connect-src-websocket-blocked</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src='../support/logTest.sub.js?logs=["blocked","violated-directive=connect-src"]'></script>
    <script src="../support/alertAssert.sub.js?alerts=[]"></script>
</head>

<body>
    <script>
        window.addEventListener('securitypolicyviolation', function(e) {
            log("violated-directive=" + e.violatedDirective);
        });

        var ws = new WebSocket("ws://{{domains[www1]}}:{{ports[http][0]}}/echo");
        ws.onerror = () => {
            assert_equals(ws.readyState, WebSocket.CLOSED, "WebSocket should be closed");
            log("blocked");
        }
    </script>
    <div id="log"></div>
</body>

</html>
